ChatBot-Langchain/requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
pandas
langgraph

#for tools
duckduckgo-search
ddgs

#for gemini
langchain-google-genai

#for async programming
aioconsole

#for UI
streamlit